requirements.txt
python-dotenv
langchain
openai
streamlit
streamlit-chat
faiss-cpu
pypdf
tiktoken
tabulate
unstructured
pdf2image
python-magic-bin
sentence_transformers 
InstructorEmbedding


gpt4all==0.2.3
chromadb==0.3.23
llama-cpp-python==0.1.50
urllib3==2.0.2
pdfminer.six==20221105
unstructured==0.6.6
extract-msg==0.41.1
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
tqdm==4.65.0
# Install torch, torchvision, torchaudio according to 
# instructions fround on pytorch.org
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu117
